requirements.txt
streamlit==1.37.1
openai>=1.37.0
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
